Change percent range from float [0-1] to int [0-100]

+++ Timer.h
@@ -15,8 +15,8 @@ class Timer {
     bool isExpired();                     // check if the time has run out
     unsigned long getValue();             // how many milliseconds that have passed since the start of the timer
     unsigned long getInverseValue();      // how many milliseconds the timer has until finished
-    float getPercentValue();              // how much time has passed as a percentage of the interval
-    float getInversePercentValue();       // how much time is left as a percentage of the interval
+    int getPercentValue();                // how much time has passed as a percentage of the interval
+    int getInversePercentValue();         // how much time is left as a percentage of the interval
   private:
     unsigned long _timeout;   // how long this timer should run for
     unsigned long _waypoint;  // the point in time the timer was started or reset
